+++ pkg/gather/gatherer/certificate_signing_requests_test.go
@@ -1,4 +1,4 @@
-package clusterconfig
+package gatherer
 
 import (
 	"encoding/json"
