+++ pkg/gather/gatherer/clusterconfig_test.go
@@ -1,4 +1,4 @@
-package clusterconfig
+package gatherer
 
 import (
 	"context"
