+++ pkg/gather/gatherer/infrastructures.go
@@ -1,0 +1,55 @@
+package gatherer
+
+import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+
+	configv1 "github.com/openshift/api/config/v1"
+	_ "k8s.io/apimachinery/pkg/runtime/serializer/yaml"
+
+	"github.com/openshift/insights-operator/pkg/record"
+)
+
+// GatherClusterInfrastructure fetches the cluster Infrastructure - the Infrastructure with name cluster.
+//
+// The Kubernetes api https://github.com/openshift/client-go/blob/master/config/clientset/versioned/typed/config/v1/infrastructure.go#L50
+// Response see https://docs.openshift.com/container-platform/4.3/rest_api/index.html#infrastructure-v1-config-openshift-io
+//
+// Location in archive: config/infrastructure/
+// See: docs/insights-archive-sample/config/infrastructure
+func GatherClusterInfrastructure(i *Gatherer) func() ([]record.Record, []error) {
+	return func() ([]record.Record, []error) {
+		config, err := i.client.Infrastructures().Get(i.ctx, "cluster", metav1.GetOptions{})
+		if errors.IsNotFound(err) {
+			return nil, nil
+		}
+		if err != nil {
+			return nil, []error{err}
+		}
+		return []record.Record{{Name: "config/infrastructure", Item: InfrastructureAnonymizer{config}}}, nil
+	}
+}
+
+// InfrastructureAnonymizer anonymizes infrastructure
+type InfrastructureAnonymizer struct{ *configv1.Infrastructure }
+
+// Marshal serializes Infrastructure with anonymization
+func (a InfrastructureAnonymizer) Marshal(_ context.Context) ([]byte, error) {
+	return runtime.Encode(openshiftSerializer, anonymizeInfrastructure(a.Infrastructure))
+}
+
+// GetExtension returns extension for anonymized infra objects
+func (a InfrastructureAnonymizer) GetExtension() string {
+	return "json"
+}
+
+func anonymizeInfrastructure(config *configv1.Infrastructure) *configv1.Infrastructure {
+	config.Status.APIServerURL = anonymizeURL(config.Status.APIServerURL)
+	config.Status.EtcdDiscoveryDomain = anonymizeURL(config.Status.EtcdDiscoveryDomain)
+	config.Status.InfrastructureName = anonymizeURL(config.Status.InfrastructureName)
+	config.Status.APIServerInternalURL = anonymizeURL(config.Status.APIServerInternalURL)
+	return config
+}
