+++ pkg/gather/gatherer/testdata/clusteroperators.json
