+++ pkg/gather/gatherer/testdata/configmaps.json
