+++ pkg/gather/gatherer/testdata/csr_appr_anon.json
