+++ pkg/gather/gatherer/testdata/installplan_openshift.yaml
