+++ pkg/gather/gatherer/testdata/nodes.json
